Result: Fix ToString_SourceIsFailureDefault test
